Refaktoryzacja

+++ Journal/AddGradeDialog.xaml.cs
@@ -1,16 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
 
 namespace Journal
 {
@@ -36,7 +25,7 @@ namespace Journal
         }
         public int InsertedGrade
         {
-            get { return Convert.ToInt16(NewGrade.Text); }
+            get { return Convert.ToInt32(NewGrade.Text); }
         }
     }
 }
